requirements.txt
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
google-genai>=1.28.0
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
sift-stack-py>=0.8.2
sqlalchemy>=2.0.42
werkzeug>=3.1.3